+++ src/ska_dataproduct_api/components/metadata/metadata.py
@@ -14,6 +14,7 @@ Functions:
 
 import datetime
 import hashlib
+import json
 import logging
 import pathlib
 import uuid
@@ -21,6 +22,8 @@ import uuid
 import yaml
 
 logger = logging.getLogger(__name__)
+
+# pylint: disable=too-many-instance-attributes
 
 
 class DataProductMetadata:
@@ -40,10 +43,11 @@ class DataProductMetadata:
         self.metadata_dict: dict = None
         self.date_created: str = None
         self.object_id: str = None
-        self.uuid: uuid.UUID = None
+        self.data_product_uuid: uuid.UUID = None
         self.execution_block: str = None
+        self.metadata_dict_hash: str = None
 
-    def get_execution_block_id(self, metadata_dict) -> str | None:
+    def get_execution_block_id(self, metadata_dict: dict) -> str | None:
         """Retrieves the execution block ID from the given metadata dictionary.
 
         Args:
@@ -51,26 +55,44 @@ class DataProductMetadata:
 
         Returns:
             str | None: The execution block ID, or None if not found.
+
+        Raises:
+            KeyError: If the `execution_block` key is not found in the metadata dictionary.
         """
+
         try:
             return metadata_dict["execution_block"]
-        except Exception as error:  # pylint: disable=broad-exception-caught
+        except KeyError as error:
             logger.error("execution_block value key not found in data product, error: %s", error)
-            return None
+            raise error
 
     def derive_uuid(self, execution_block_id: str, file_path: pathlib.Path) -> uuid.UUID:
         """Derives a UUID from an execution block ID and file path.
 
         Args:
-            execution_block_id: The execution block ID.
-            file_path: The file path.
+            execution_block_id (str): The execution block ID.
+            file_path (pathlib.Path): The file path.
 
         Returns:
-            str | None: The derived UUID, or None if not found.
+            uuid.UUID: The derived UUID.
+
+        Raises:
+            ValueError: If the execution block ID is None.
+            ValueError: If the UUID cannot be created.
         """
         if execution_block_id is None:
-            logger.error("No execution_block_id, so the UUID is not set")
-            return None
+            logger.error("Execution block ID cannot be None.")
+            raise ValueError("Execution block ID cannot be None.")
+
+        # I am using a combination of the execution_block_id and file_path of the data products
+        # to derive a uuid. The file path would be unique and consistent for the initial use as the
+        # DPD only loads data products from one PV, but as soon as the DLM adds data products
+        # to the dashboard, there might not be a direct reference to a file on disk.
+
+        # This is also not envisioned to be the source of the global UUID of data products, it
+        # should only be the DLM creating and assigning these uuid's, but for internal use for the
+        # DPD, to cater for data products that is loaded by the DPD from the PV or for sub
+        # products, we assign a local uuid here.
 
         try:
             combined_string = f"{execution_block_id}:{str(file_path)}"
@@ -79,9 +101,21 @@ class DataProductMetadata:
 {hash_value[16:20]}-{hash_value[20:32]}"
             uuid_value = uuid.UUID(formatted_hash)
             return uuid_value
-        except Exception as error:  # pylint: disable=broad-exception-caught
-            logger.error("Failed to create uuid, error: %s", error)
-            return None
+        except ValueError as error:
+            logger.error("Failed to create UUID: %s", error)
+            raise error
+
+    def calculate_metadata_hash(self, metadata_file_json: dict) -> str:
+        """Calculates a SHA256 hash of the given metadata JSON.
+
+        Args:
+            metadata_file_json (dict): The metadata JSON to be hashed.
+
+        Returns:
+            str: The SHA256 hash of the metadata JSON.
+        """
+
+        return hashlib.sha256(json.dumps(metadata_file_json).encode("utf-8")).hexdigest()
 
     def load_yaml_file(self, file_path: pathlib.Path) -> None:
         """
@@ -100,10 +134,6 @@ class DataProductMetadata:
         try:
             with open(self.data_product_metadata_file_path, "r", encoding="utf-8") as file:
                 self.metadata_dict = yaml.safe_load(file)
-                self.execution_block = self.get_execution_block_id(self.metadata_dict)
-                self.uuid = self.derive_uuid(
-                    execution_block_id=self.execution_block, file_path=self.data_product_file_path
-                )
         except FileNotFoundError as error:
             raise FileNotFoundError(
                 f"Metadata file not found: {self.data_product_metadata_file_path}"
@@ -112,6 +142,12 @@ class DataProductMetadata:
             raise yaml.YAMLError(
                 f"Error parsing YAML file: {self.data_product_metadata_file_path}"
             ) from error
+
+        self.execution_block = self.get_execution_block_id(self.metadata_dict)
+        self.metadata_dict_hash = self.calculate_metadata_hash(self.metadata_dict)
+        self.data_product_uuid = self.derive_uuid(
+            execution_block_id=self.execution_block, file_path=self.data_product_file_path
+        )
 
     def load_metadata_from_yaml_file(self, file_path: pathlib.Path) -> dict[str, any]:
         """
@@ -123,7 +159,12 @@ class DataProductMetadata:
         Returns:
             dict[str, any]: Loaded metadata as a dictionary.
         """
-        self.load_yaml_file(file_path=file_path)
+        try:
+            self.load_yaml_file(file_path=file_path)
+        except Exception as error:
+            logger.error("Failed to load metadata, error: %s", error)
+            raise error
+
         self.append_metadata()
         return self.metadata_dict
 
@@ -139,7 +180,8 @@ class DataProductMetadata:
         """
         self.metadata_dict = metadata
         self.execution_block = self.get_execution_block_id(self.metadata_dict)
-        self.uuid = self.derive_uuid(
+        self.metadata_dict_hash = self.calculate_metadata_hash(self.metadata_dict)
+        self.data_product_uuid = self.derive_uuid(
             execution_block_id=self.execution_block, file_path=self.data_product_file_path
         )
         self.append_metadata()
@@ -191,7 +233,7 @@ class DataProductMetadata:
                 "date_created": self.date_created,
                 "dataproduct_file": str(self.data_product_file_path),
                 "metadata_file": str(self.data_product_metadata_file_path),
-                "uuid": str(self.uuid),
+                "uuid": str(self.data_product_uuid),
             }
         )