+++ src/ska_dataproduct_api/components/search/in_memory/in_memory_search.py
@@ -39,7 +39,7 @@ class InMemoryDataproductSearch(MetadataSearchStore):
         muiDataGridInstance.flattened_list_of_dataproducts_metadata.clear()
         self.load_metadata_from_store()
 
-    def insert_metadata_in_search_store(self, metadata_dict: dict):
+    def insert_metadata_in_search_store(self, metadata_dict: dict) -> None:
         """This method loads the metadata file of a data product, creates a
         list of keys used in it, and then adds it to the flattened_list_of_dataproducts_metadata"""
         # generate a list of keys from this object
@@ -52,7 +52,9 @@ class InMemoryDataproductSearch(MetadataSearchStore):
             list_of_dict=muiDataGridInstance.flattened_list_of_dataproducts_metadata
         )
 
-        self.number_of_dataproducts = self.number_of_dataproducts + 1
+        self.number_of_dataproducts = len(
+            muiDataGridInstance.flattened_list_of_dataproducts_metadata
+        )
 
     def sort_list_of_dict(
         self, list_of_dict: list[dict], key: str = "date_created", reverse: bool = True